[Refactor] Refactor all utils in models
