+++ mmflow/models/decoders/liteflownet_decoder.py
@@ -11,8 +11,8 @@ from mmcv.runner import BaseModule
 
 from mmflow.ops import build_operators
 from ..builder import DECODERS, build_loss
+from ..utils import CorrBlock
 from .base_decoder import BaseDecoder
-from .utils import CorrBlock
 
 
 class Upsample(nn.Module):
