+++ mmflow/models/utils/__init__.py
@@ -1,5 +1,12 @@
 # Copyright (c) OpenMMLab. All rights reserved.
+from .basic_encoder import BasicConvBlock, BasicEncoder
+from .correlation_block import CorrBlock
+from .densenet import BasicDenseBlock, DenseLayer
 from .estimators_link import BasicLink, LinkOutput
 from .res_layer import BasicBlock, Bottleneck, ResLayer
 
-__all__ = ['ResLayer', 'BasicBlock', 'Bottleneck', 'BasicLink', 'LinkOutput']
+__all__ = [
+    'ResLayer', 'BasicBlock', 'Bottleneck', 'BasicLink', 'LinkOutput',
+    'DenseLayer', 'BasicDenseBlock', 'BasicEncoder', 'BasicConvBlock',
+    'CorrBlock'
+]
