+++ tests/test_models/test_flow_estimator.py
@@ -12,15 +12,12 @@ from mmflow.models.flow_estimators.base import FlowEstimator
 @pytest.mark.skipif(not torch.cuda.is_available(), reason='CUDA not available')
 @pytest.mark.parametrize('cfg_file', [
     '../../configs/_base_/models/pwcnet.py',
-    '../../configs/_base_/models/raft.py',
     '../../configs/_base_/models/maskflownets.py',
     '../../configs/_base_/models/maskflownet.py',
     '../../configs/_base_/models/flownetc.py',
-    '../../configs/_base_/models/flownets.py',
     '../../configs/_base_/models/liteflownet/liteflownet.py',
     '../../configs/_base_/models/flownet2/flownet2cs.py',
     '../../configs/_base_/models/flownet2/flownet2css.py',
-    '../../configs/_base_/models/flownet2/flownet2sd.py',
     '../../configs/_base_/models/flownet2/flownet2.py'
 ])
 def test_flow_estimator(cfg_file):
@@ -36,6 +33,32 @@ def test_flow_estimator(cfg_file):
     imgs = torch.randn(1, 6, 64, 64).cuda()
     flow_gt = torch.randn(1, 2, 64, 64).cuda()
     valid = torch.ones((1, 64, 64)).cuda()
+    # test forward_test out
+    assert isinstance(estimator(imgs, test_mode=True), list)
+
+    losses = estimator(imgs, flow_gt, valid, test_mode=False)
+    loss, _ = estimator._parse_losses(losses)
+
+    assert float(loss.item()) > 0
+
+
+@pytest.mark.parametrize('cfg_file', [
+    '../../configs/_base_/models/raft.py',
+    '../../configs/_base_/models/flownets.py',
+    '../../configs/_base_/models/flownet2/flownet2sd.py'
+])
+def test_flow_estimator_without_cuda(cfg_file):
+    # BaseFlowEstimator has abstract method
+    with pytest.raises(TypeError):
+        FlowEstimator(init_cfg=None)
+
+    cfg_file = osp.join(osp.dirname(__file__), cfg_file)
+    cfg = Config.fromfile(cfg_file)
+
+    estimator = build_flow_estimator(cfg.model)
+    imgs = torch.randn(1, 6, 64, 64)
+    flow_gt = torch.randn(1, 2, 64, 64)
+    valid = torch.ones((1, 64, 64))
     # test forward_test out
     assert isinstance(estimator(imgs, test_mode=True), list)
